few changes

+++ dotfilesInstaller.sh
@@ -5,7 +5,7 @@ ln -s ~/dotfiles/Brewfile ~/Brewfile
 ln -s ~/dotfiles/.gitconfig ~/.gitconfig
 ln -s ~/dotfiles/.gitignore ~/.gitignore
 
-rm ~/Library/Preferences/com.googlecode.iterm2.plist
+test -e ~/Library/Preferences/com.googlecode.iterm2.plist && rm ~/Library/Preferences/com.googlecode.iterm2.plist
 ln -s ~/dotfiles/.iterm2/com.googlecode.iterm2.plist ~/Library/Preferences/com.googlecode.iterm2.plist
 ln -s ~/dotfiles/.iterm2/ ~/.iterm2
 
@@ -14,13 +14,6 @@ ln -s ~/dotfiles/.zshrc ~/.zshrc
 
 ln -s ~/dotfiles/.vim/ ~/.vim
 ln -s ~/dotfiles/.vimrc ~/.vimrc
-
-    # ln -s ~/dotfiles/adb-peco/bin/adb_peco.sh ~/dotfiles/bin/adb_peco.sh
-    # ln -s ~/dotfiles/adb-peco/bin/adbp ~/dotfiles/bin/adbp
-    # ln -s ~/dotfiles/adb-peco/bin/pidcatp ~/dotfiles/bin/pidcatp
-
-    # ln -s ~/dotfiles/ADB-Tools/bin/adb-screencap ~/dotfiles/bin/adb-screencap
-    # ln -s ~/dotfiles/ADB-Tools/bin/adb-screenrecord ~/dotfiles/bin/adb-screenrecord
 
 ln -s ~/dotfiles/bin/ ~/bin
